+++ web/src/router/index.ts
@@ -12,11 +12,6 @@ const routes: Array<RouteRecordRaw> = [
     component: Home
   },
   {
-    path: '/login',
-    name: 'Login',
-    component: Login
-  },
-  {
     path: '/register',
     name: 'Register',
     component: Register
